+++ src/SimpleCache/CacheExeption.php
@@ -1,7 +1,14 @@
 <?php
+/**
+ * CacheException
+ *
+ * @package Nofuzz
+ */
+#################################################################################################################################
 
 namespace Nofuzz\SimpleCache;
 
-class CacheException extends \RuntimeException
+class CacheException extends \Exception
 {
+
 }